+++ src/rng/rand_max2.c
@@ -1,6 +1,6 @@
 #include "mt19937.h"
 
-u32 Mt19937RandMax2(u16 max)
+u16 Mt19937RandMax2(u16 max)
 {
     u32 result = Mt19937Next2();
     result &= 0x7FFFu;
